+++ src/main/java/com/cisowski/schoolmanagement/timetable/attendance/service/AttendanceService.java
@@ -2,11 +2,11 @@ package com.cisowski.schoolmanagement.timetable.attendance.service;
 
 import com.cisowski.schoolmanagement.timetable.attendance.model.AttendanceSummaryResponse;
 import com.cisowski.schoolmanagement.timetable.attendance.model.MarkAttendanceRequest;
-import com.cisowski.schoolmanagement.timetable.schedule.model.ScheduleEntity;
+import com.cisowski.schoolmanagement.timetable.scheduleOccurrence.model.ScheduleOccurrenceEntity;
 
 import java.util.List;
 
 public interface AttendanceService {
-    void initializeAttendances(ScheduleEntity schedule);
-    List<AttendanceSummaryResponse> setAttendanceAbsenceStatusForStudents(Integer scheduleId, MarkAttendanceRequest request);
+    void initializeAttendances(ScheduleOccurrenceEntity scheduleOccurrence);
+    List<AttendanceSummaryResponse> setAttendanceAbsenceStatusForStudents(Long scheduleOccurrenceId, MarkAttendanceRequest request);
 }