+++ src/main/java/com/cisowski/schoolmanagement/timetable/schedule/model/ScheduleStatus.java
@@ -6,7 +6,5 @@ public enum ScheduleStatus {
     CANCELLED,
     RESCHEDULED,
     DELETED,
-    COMPLETED,
-    UPDATED,
-    ONGOING;
+    UPDATED;
 }